pyproject.toml
[project]
name = "FumeTune"
version = "1.0.0"
description = "Unlock the ultimate music experience for YOUR Discord community with FumeTune!"
license = "AGPL-3.0"
authors = [
    { name="Sayan Bhattacharyya", email="sayan@sn1f3rt.dev" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.2.0",
    "async-timeout>=5.0.1",
    "azapi>=3.0.8",
    "better-ipc",
    "discord-ext-menus",
    "discord-ext-menus-views",
    "discord-py[speed,voice]>=2.4.0",
    "topggpy>=1.4.0",
    "wavelink>=3.4.1",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1",
    "ruff>=0.8.3",
]

[tool.uv.sources]
better-ipc = { git = "https://github.com/MiroslavRosenov/better-ipc", rev = "efb145d9dabc0a90e4c6e32be08bd03ae4257249" }
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus" }
discord-ext-menus-views = { git = "https://github.com/FumeStop/discord-ext-menus-views" }
